(0.0.1.24)
Hey! It's been awhile, so I'm not really able to detail the exact changes for this commit, however, I've worked more on the job system, and commutes.

Also, I started outlining a new Assets System, that would allow for the very abstract and complex dynamics of stuff like vehicle ownership and availability, while also being flexible enough for use on other systems.